fix(commands): keep headless key parsing OpenTUI-free
